Refactor diff methods into a more cohesive structure.
